I lied about release.
I had a one time app breaking crash caused by the D2XX library.  I don't
know why it caused it, and have been unable to reproduce it.

As such, I plan to replace it with the android USB Serial library.  I
may need to fork that library to be sure that the DTR line is off on
startup.

This commit  fixes a typo with the power delay, and changes the
subchannel member of tuneinfo so that it is not final and can be
changed, as that member is only used for tracking or tuning to the
subchannel, it is not received from a tune replay.